First class functions in IR, inlined eagerly
Summary: This is a continuation of #18367 that rebases it and cleans up
some of the issues that were popping up.

Notes:

1. removed special handling for Python 2 print statements. Its
implementation was buggy (treated inlined tuples as literal tuples),
and leaving it out just means that in Python 2 printing multiple values
will print different in script and python. This is ok, because it is
minor formatting differences and only in Python 2.

2. --accepted internal loss of ONNX type information. We are only
required to have correct input and output types for ONNX not for
intermediates.

gh-metadata: pytorch pytorch 21052 gh/zdevito/40/head